Pull request #4162: coverity: fix for stream and hash
Merge in SNORT/snort3 from ~RSHAFIQ/snort3:stream_coverity to master

Squashed commit of the following:

commit ebffbe09dda5a5733f86c683f16347716a0a51ce
Author: rshafiq <rshafiq@cisco.com>
Date:   Tue Dec 19 14:15:41 2023 -0500

    coverity: fix for stream and hash

+++ src/service_inspectors/netflow/netflow.cc
@@ -123,8 +123,8 @@ static void publish_netflow_event(const Packet* p, const NetFlowRule* match, Net
     // LAST_PKT_SECOND - if these aren't set, assume the current wire pkt time
     if (!record.first_pkt_second or !record.last_pkt_second)
     {
-        record.first_pkt_second = packet_time();
-        record.last_pkt_second = packet_time();
+        record.first_pkt_second = static_cast<uint32_t>(packet_time());
+        record.last_pkt_second = static_cast<uint32_t>(packet_time());
     }
 
     NetFlowEvent event(p, &record, match->create_host, match->create_service, swapped, serviceID);
